Moved text formatting classes to their own file
Moved the classes which contained the codes for text colours and formatting to a file of their own
