+++ main.py
@@ -3,7 +3,7 @@
 
 try:
     from shellexceptions import GenericException, ImportException
-    import os, pwd, datetime, platform, getopt
+    import os, pwd, datetime, platform, getopt, textcodes
     import ls, cd, mkdir, rmdir, cp, rm, printworkingdir
     import exittheshell, cleartheshell
 except ImportError as e:  # If any module failed to be imported
@@ -21,46 +21,12 @@ except ImportError as e:  # If any module failed to be imported
                 exit()
             else:
                 exit()
-    except Exception:
-        exit()
+    except:
+        exit_from_exception()
 
 
-class TerminalColours(object):
-    """ Class containing the colour codes for the terminal """
-    def __init__(self):
-        # Colour codes
-        self._blue_colour = "\033[94m"
-        self._cyan_colour = "\033[96m"
-        self._darker_cyan_colour = "\033[36m"
-        self._green_colour = "\033[92m"
-        self._yellow_colour = "\033[93m"
-        self._purple_colour = "\033[95m"
-        self._red_colour = "\033[91m"
-
-    def get_colour_code(self, colour_name):
-        """ Fetches the colour code for the colour name passed """
-        return getattr(self, "_"+colour_name+"_colour")
-
-
-class TerminalFormatting(object):
-    """ Class containing the formatting codes for the terminal """
-    def __init__(self):
-        # Formatting codes
-        self._bold_formatting = "\033[1m"
-        self._end_of_colours = "\033[0m"
-        self._underline_formatting = "\033[4m"
-        self._blink_formatting = "\033[5m"
-
-    def get_formatting_code(self, formatting_code_name):
-        """ Fetches the formatting code for the formmating code name passed """
-        if formatting_code_name == "end":
-            return self.__getattribute__("_end_of_colours")
-        else:
-            return getattr(self, "_"+formatting_code_name+"_formatting")
-
-
-TerminalColourInstance = TerminalColours()
-TerminalFormattingInstance = TerminalFormatting()
+TerminalColourInstance = textcodes.TerminalColours()
+TerminalFormattingInstance = textcodes.TerminalFormatting()
 red_warning_text = TerminalColourInstance.get_colour_code("red")
 red_warning_text += TerminalFormattingInstance.get_formatting_code("bold")
 red_warning_text += TerminalFormattingInstance.get_formatting_code("blink")
